refactor(sim): extract shared IPC into sim_runtime.ipc
Extrahiert den in run_twitter_simulation.py und run_reddit_simulation.py
duplizierten IPC-Block (Konstanten, CommandType, IPCHandler) in das neue
Paket backend/scripts/sim_runtime/ipc.py. Beide Runner importieren nun
die gemeinsame Implementierung; die IPCHandler-Konstruktion injiziert
db_filename, interview_action_type (ActionType.INTERVIEW) und
manual_action_cls (ManualAction) als Keyword-Argumente.

Verhaltensidentisch: Datei-IPC-Layout, Response-Format, Command-Dedup
(seen_command_ids) und die _execute_command-Ruckgabematrix bleiben
unverandert. Oasis-Symbole werden injiziert statt auf Modulebene importiert,
damit sim_runtime.ipc ohne torch-abhangige Oasis-Laufzeit importierbar bleibt
und im Test-Tree charakterisierbar ist.

Erhaltungsanker: _sim_common bleibt eigenstandiges Modul (importlib-Tests);
create_model/ModelFactory bleiben Attribute von run_parallel_simulation
(test_oasis_provider_dispatch); die LLM_MAX_OUTPUT_TOKENS/--num-agents/
--num-rounds-Literale bleiben in den Entry-Points.
